requirements.txt
# Mind the Gap - Advanced Dependencies

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Reddit API
praw>=7.7.0
prawcore>=2.3.0

# Deep Learning & Transformers
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2

# NLP Preprocessing
nltk>=3.8.1
spacy>=3.7.0
regex>=2023.10.3

# Topic Modeling
bertopic>=0.16.0
scikit-learn>=1.3.0
umap-learn>=0.5.4
hdbscan>=0.8.33

# Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.2

# Explainable AI
shap>=0.43.0
lime>=0.2.0.1

# Time Series
prophet>=1.1.5

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
schedule>=1.2.0
pytz>=2023.3

# Database (optional)
sqlalchemy>=2.0.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0